requirements.txt
build==1.4.2
certifi==2026.2.25
confluent-kafka==2.13.2
dnspython==2.8.0
grpcio==1.80.0
grpcio-tools==1.80.0
iniconfig==2.3.0
msgpack==1.1.2
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pluggy==1.6.0
protobuf==6.33.6
Pygments==2.20.0
pymongo==4.16.0
pyproject_hooks==1.2.0
pytest==9.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
setuptools==82.0.1
six==1.17.0
TA-Lib==0.6.8
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
